src/main/java/com/springapp/mvc/controller/FinanceController.java
package com.springapp.mvc.controller;

import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.servlet.ModelAndView;

@Controller
public class FinanceController {
    @RequestMapping(value = "/", method = RequestMethod.GET)
    public ModelAndView showIndex() {

        ModelAndView index = new ModelAndView("index");
        index.getModel().put("message", "ThoughtWorks index page");
        return index;
    }

}
